Created the TeamsHandler class and used it in index.ts
